requirements.txt
Flask
Pillow
easyocr
flasgger
uvicorn
gunicorn
MarkupSafe==0.15
Pillow==2.0.0
argparse==1.2.1
numpy==1.7.1
pycups==1.9.63
pycurl==7.19.0
pygobject==3.10.0
pysmbc==1.0.13
pyxdg==0.25
requests==1.2.3
six==1.3.0